Implement a server URL selector and refactor the Home screen UI components.
*   Extract `SharingHeroSection` and `AdvancedSettingsSection` from `HomeScreen.kt` into dedicated component files.
*   Implement `ServerUrlSelector` component to allow users to choose between available IP addresses (including IPv6 support) for the sharing link and QR code.
*   Update `HomeState` and `HomeViewModel` to manage the selected server URL and its selector's expanded state.
*   Refactor `SharingHeroSection` to use the new `ServerUrlSelector` instead of a static text display.
*   Fix IPv6 address formatting in `ServerUrlProvider` for generated URLs.
*   Update the version 9 changelog to reflect that the IP address for the QR code is now selectable.
