+++ app/src/main/java/com/defname/localshare/ui/screens/home/HomeScreen.kt
@@ -11,12 +11,7 @@ import android.os.Build
 import android.text.format.Formatter
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -33,24 +28,16 @@ import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Cancel
 import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.ExpandLess
-import androidx.compose.material.icons.filled.ExpandMore
 import androidx.compose.material.icons.filled.List
 import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.icons.filled.QrCode2
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Share
 import androidx.compose.material.icons.filled.Undo
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalMinimumInteractiveComponentSize
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
@@ -59,27 +46,22 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.defname.localshare.ui.components.HintCard
 import com.defname.localshare.ui.components.LogListEntry
 import com.defname.localshare.ui.components.QrCodeDialog
+import com.defname.localshare.ui.screens.home.components.AdvancedSettingsSection
 import com.defname.localshare.ui.screens.home.components.FileCarousel
-import com.defname.localshare.ui.screens.home.components.IpAddressSelector
+import com.defname.localshare.ui.screens.home.components.SharingHeroSection
 import com.defname.localshare.ui.screens.home.components.StartServerButton
 import com.defname.localshare.ui.theme.LocalShareTheme
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
 import org.koin.androidx.compose.koinViewModel
 
 fun shareText(context: Context, text: String) {
@@ -328,170 +310,6 @@ fun HomeScreen(
                 }
             }
             Spacer(modifier = Modifier.height(80.dp))
-        }
-    }
-}
-
-@Composable
-fun SharingHeroSection(state: HomeState, viewModel: HomeViewModel, context: Context) {
-    val isRunning = state.isRunning
-    val primaryUrl = state.primaryServerUrl
-
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = if (isRunning) MaterialTheme.colorScheme.primaryContainer
-            else MaterialTheme.colorScheme.surfaceVariant
-        ),
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = if (isRunning) 4.dp else 0.dp
-        ),
-        border = if (!isRunning) BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant) else null
-    ) {
-        Column(
-            modifier = Modifier.padding(24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Icon(
-                imageVector = if (isRunning) Icons.Default.Share else Icons.Default.Cancel,
-                contentDescription = null,
-                modifier = Modifier.size(48.dp),
-                tint = if (isRunning) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            
-            Spacer(modifier = Modifier.height(16.dp))
-
-            Text(
-                text = if (isRunning) "Sharing is Active" else "Sharing is Off",
-                style = MaterialTheme.typography.headlineSmall,
-                fontWeight = FontWeight.Bold,
-                color = if (isRunning) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurfaceVariant
-            )
-
-            if (isRunning && primaryUrl != null) {
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = "Anyone on your Wi-Fi can access files at:",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onPrimaryContainer.copy(alpha = 0.8f),
-                    textAlign = TextAlign.Center
-                )
-                Spacer(modifier = Modifier.height(12.dp))
-
-                Card(
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.1f)
-                    ),
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-
-                    Text(
-                        text = primaryUrl,
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Bold,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.fillMaxWidth().padding(12.dp),
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(24.dp))
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Button(
-                        onClick = { shareText(context, primaryUrl) },
-                        modifier = Modifier.weight(1f),
-                        contentPadding = PaddingValues(vertical = 12.dp)
-                    ) {
-                        Icon(Icons.Default.Share, contentDescription = null, modifier = Modifier.size(18.dp))
-                        Spacer(Modifier.width(8.dp))
-                        Text("Share Link")
-                    }
-                    Button(
-                        onClick = { viewModel.onOpenQrCodeDialog(primaryUrl) },
-                        modifier = Modifier.weight(1f),
-                        colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.secondary),
-                        contentPadding = PaddingValues(vertical = 12.dp)
-                    ) {
-                        Icon(Icons.Default.QrCode2, contentDescription = null, modifier = Modifier.size(18.dp))
-                        Spacer(Modifier.width(8.dp))
-                        Text("QR Code")
-                    }
-                }
-            } else if (!isRunning) {
-                Text(
-                    "Start sharing to make your files available to other devices on this network.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    textAlign = TextAlign.Center,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun AdvancedSettingsSection(
-    state: HomeState,
-    viewModel: HomeViewModel,
-    scope: CoroutineScope
-) {
-    var expanded by remember { mutableStateOf(false) }
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .animateContentSize()
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { expanded = !expanded }
-                .padding(vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                "Technical Details",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                modifier = Modifier.weight(1f)
-            )
-            Icon(
-                imageVector = if (expanded) Icons.Default.ExpandLess else Icons.Default.ExpandMore,
-                contentDescription = null
-            )
-        }
-
-        if (expanded) {
-            Column(modifier = Modifier.padding(top = 8.dp, bottom = 16.dp)) {
-                OutlinedTextField(
-                    value = state.token,
-                    label = { Text("Security Token") },
-                    singleLine = true,
-                    onValueChange = { scope.launch { viewModel.onTokenChange(it) } },
-                    modifier = Modifier.fillMaxWidth(),
-                    trailingIcon = {
-                        IconButton(onClick = { scope.launch { viewModel.onRandomTokenClick() } }) {
-                            Icon(Icons.Default.Refresh, contentDescription = "Generate New")
-                        }
-                    }
-                )
-
-                Spacer(modifier = Modifier.height(16.dp))
-                
-                IpAddressSelector(
-                    addresses = state.localIpAddresses,
-                    selectedAddress = state.selectedIpAddress,
-                    isSelectedAddressValid = state.isSelectedIpAddressValid,
-                    expanded = state.ipAddressSelectorExpanded,
-                    enabled = state.ipAddressSelectorEnabled,
-                    onExpandedChange = { viewModel.ipAddressSelectorExpandedChange() },
-                    onAddressSelected = { scope.launch { viewModel.onIpAddressSelected(it) } },
-                    onDismiss = { viewModel.collapseIpAddressSelector() }
-                )
-            }
         }
     }
 }
