+++ app/src/main/java/com/defname/localshare/ui/screens/home/components/ServerUrlSelector.kt
@@ -1,0 +1,94 @@
+package com.defname.localshare.ui.screens.home.components
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowDropDown
+import androidx.compose.material.icons.filled.ArrowDropUp
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExposedDropdownMenuBox
+import androidx.compose.material3.ExposedDropdownMenuDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ServerUrlSelector(
+    availableUrls: List<String> = emptyList(),
+    selectedUrl: String,
+    expanded: Boolean = false,
+    onExpandedChange: () -> Unit = {},
+    onUrlSelected: (String) -> Unit = {},
+    onDismiss: () -> Unit = {},
+) {
+    val enabled = availableUrls.size > 1
+
+    ExposedDropdownMenuBox(
+        expanded = expanded,
+        onExpandedChange = { if (enabled) onExpandedChange() },
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 8.dp)
+    ) {
+        Card(
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.1f)
+            ),
+            modifier = Modifier.fillMaxWidth().menuAnchor()
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = selectedUrl,
+                    style = MaterialTheme.typography.titleMedium,
+                    fontWeight = FontWeight.Bold,
+                    textAlign = TextAlign.Center,
+                    modifier = Modifier.weight(1.0f).padding(12.dp),
+                    color = MaterialTheme.colorScheme.primary,
+                )
+
+                if (enabled) {
+                    Icon(
+                        imageVector = if (expanded) Icons.Default.ArrowDropUp else Icons.Default.ArrowDropDown,
+                        contentDescription = null,
+                        tint = MaterialTheme.colorScheme.primary,
+                        modifier = Modifier.padding(12.dp)
+                    )
+                }
+            }
+        }
+
+        ExposedDropdownMenu(
+            expanded = expanded,
+            onDismissRequest = { onDismiss() }
+        ) {
+            // Option 1: All Networks
+            // Dynamische Optionen aus der IP-Liste
+            availableUrls.forEach { url ->
+                DropdownMenuItem(
+                    text = {
+                        Text(url, style = MaterialTheme.typography.bodyLarge)
+                    },
+                    onClick = {
+                        onUrlSelected(url)
+                        onDismiss()
+                    },
+                    contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding
+                )
+            }
+        }
+    }
+}
