+++ app/src/main/java/com/defname/localshare/ui/screens/home/components/SharingHeroSection.kt
@@ -1,0 +1,127 @@
+package com.defname.localshare.ui.screens.home.components
+
+import android.content.Context
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Cancel
+import androidx.compose.material.icons.filled.QrCode2
+import androidx.compose.material.icons.filled.Share
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.defname.localshare.ui.screens.home.HomeState
+import com.defname.localshare.ui.screens.home.HomeViewModel
+import com.defname.localshare.ui.screens.home.shareText
+
+
+@Composable
+fun SharingHeroSection(state: HomeState, viewModel: HomeViewModel, context: Context) {
+    val isRunning = state.isRunning
+    val primaryUrl = state.selectedServerUrl
+
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        colors = CardDefaults.cardColors(
+            containerColor = if (isRunning) MaterialTheme.colorScheme.primaryContainer
+            else MaterialTheme.colorScheme.surfaceVariant
+        ),
+        elevation = CardDefaults.cardElevation(
+            defaultElevation = if (isRunning) 4.dp else 0.dp
+        ),
+        border = if (!isRunning) BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant) else null
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Icon(
+                imageVector = if (isRunning) Icons.Default.Share else Icons.Default.Cancel,
+                contentDescription = null,
+                modifier = Modifier.size(48.dp),
+                tint = if (isRunning) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurfaceVariant
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Text(
+                text = if (isRunning) "Sharing is Active" else "Sharing is Off",
+                style = MaterialTheme.typography.headlineSmall,
+                fontWeight = FontWeight.Bold,
+                color = if (isRunning) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurfaceVariant
+            )
+
+            if (isRunning && primaryUrl != null) {
+                Spacer(modifier = Modifier.height(8.dp))
+                Text(
+                    text = "Anyone on your Wi-Fi can access files at:",
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.onPrimaryContainer.copy(alpha = 0.8f),
+                    textAlign = TextAlign.Center
+                )
+                Spacer(modifier = Modifier.height(12.dp))
+
+                ServerUrlSelector(
+                    availableUrls = state.serverUrls,
+                    selectedUrl = state.selectedServerUrl,
+                    expanded = state.serverUrlSelectorExpanded,
+                    onExpandedChange = viewModel::onServerUrlSelectorExpandedChange,
+                    onUrlSelected = viewModel::onServerUrlSelected,
+                    onDismiss = viewModel::onServerUrlSelectorExpandedChange
+                )
+
+                Spacer(modifier = Modifier.height(24.dp))
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    Button(
+                        onClick = { shareText(context, primaryUrl) },
+                        modifier = Modifier.weight(1f),
+                        contentPadding = PaddingValues(vertical = 12.dp)
+                    ) {
+                        Icon(Icons.Default.Share, contentDescription = null, modifier = Modifier.size(18.dp))
+                        Spacer(Modifier.width(8.dp))
+                        Text("Share Link")
+                    }
+                    Button(
+                        onClick = { viewModel.onOpenQrCodeDialog(primaryUrl) },
+                        modifier = Modifier.weight(1f),
+                        colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.secondary),
+                        contentPadding = PaddingValues(vertical = 12.dp)
+                    ) {
+                        Icon(Icons.Default.QrCode2, contentDescription = null, modifier = Modifier.size(18.dp))
+                        Spacer(Modifier.width(8.dp))
+                        Text("QR Code")
+                    }
+                }
+            } else if (!isRunning) {
+                Text(
+                    "Start sharing to make your files available to other devices on this network.",
+                    style = MaterialTheme.typography.bodyMedium,
+                    textAlign = TextAlign.Center,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+        }
+    }
+}
